Remove renault from hosted video template
So template is independent of specific content.
