+++ commercial/app/views/hosted/guardianHostedGalleryImage.scala.html
@@ -1,4 +1,4 @@
-@import common.commercial.HostedGalleryImage
+@import common.commercial.hosted.HostedGalleryImage
 @(image: HostedGalleryImage, content: String)
     <div class="js-hosted-gallery-image hosted-gallery__image hosted-tone--dark">
         <div class="hosted-gallery__image-sizer">
