+++ common/app/common/commercial/hosted/HostedVideoPage.scala
@@ -1,0 +1,60 @@
+package common.commercial.hosted
+
+import model.GuardianContentTypes._
+import model.{MetaData, SectionSummary}
+import play.api.libs.json.JsString
+
+case class HostedVideoPage(
+  sectionId: String,
+  pageUrl: String,
+  pageName: String,
+  standfirst: String,
+  logoUrl: String,
+  bannerUrl: String,
+  video: HostedVideo,
+  nextPage: Option[HostedVideoPage],
+  owner: String
+) extends HostedPage {
+
+  val pageTitle: String  = s"Advertiser content hosted by the Guardian: ${video.title} - video"
+
+  override val metadata: MetaData = {
+    val toneId = "tone/hosted-content"
+    val toneName = "Hosted content"
+    val keywordId = s"$sectionId/$sectionId"
+    val keywordName = sectionId
+    MetaData.make(
+      id = s"commercial/advertiser-content/$sectionId/$pageName",
+      webTitle = pageTitle,
+      section = Some(SectionSummary.fromId(sectionId)),
+      contentType = Hosted,
+      analyticsName = s"GFE:$sectionId:$Hosted:$pageName",
+      description = Some(standfirst),
+      javascriptConfigOverrides = Map(
+        "keywordIds" -> JsString(keywordId),
+        "keywords" -> JsString(keywordName),
+        "toneIds" -> JsString(toneId),
+        "tones" -> JsString(toneName)
+      ),
+      opengraphPropertiesOverrides = Map(
+        "og:url" -> pageUrl,
+        "og:title" -> pageTitle,
+        "og:description" ->
+        s"ADVERTISER CONTENT FROM ${owner.toUpperCase} HOSTED BY THE GUARDIAN | $standfirst",
+        "og:image" -> video.posterUrl,
+        "fb:app_id" -> "180444840287"
+      )
+    )
+  }
+}
+
+case class HostedVideo(
+  mediaId: String,
+  title: String,
+  duration: Int,
+  posterUrl: String,
+  srcUrlMp4: String,
+  srcUrlWebm: String,
+  srcUrlOgg: String,
+  srcM3u8: String
+)
