+++ common/app/common/commercial/hosted/RenaultHostedPages.scala
@@ -1,4 +1,4 @@
-package common.commercial
+package common.commercial.hosted
 
 import conf.Static
 
@@ -8,12 +8,12 @@ object RenaultHostedPages {
   private val episode1PageName = "design-competition-episode1"
   private val episode2PageName = "design-competition-episode2"
 
-  private val teaser: HostedVideoPage = HostedVideoPage(
+  private val teaserWithoutNextPage: HostedVideoPage = HostedVideoPage(
+    sectionId = "renault-car-of-the-future",
     pageUrl = "https://www.theguardian.com/commercial/advertiser-content/renault-car-of-the-future/design-competition-teaser",
     pageName = teaserPageName,
     standfirst = "Who better to dream up the cars of tomorrow than the people who'll be buying them? Students at Central St Martins are working with Renault to design the interior for cars that will drive themselves. Watch this short video to find out more about the project, and visit this page again soon to catch up on the students' progress",
     logoUrl = Static("images/commercial/logo_renault.jpg"),
-    zoeUrl = Static("images/commercial/ren_zoe.jpg"),
     bannerUrl = Static("images/commercial/ren_commercial_banner.jpg"),
     video = HostedVideo(
       mediaId = "renault-car-of-the-future",
@@ -25,15 +25,16 @@ object RenaultHostedPages {
       srcUrlOgg = "https://cdn.theguardian.tv/interactive/mp4/1080/2016/05/17/160516GlabsTestSD-3_hi.ogv",
       srcM3u8 = "https://cdn.theguardian.tv/interactive/2016/05/17/HLS/160516GlabsTestSD.m3u8"
     ),
-    nextPageName = episode1PageName
+    nextPage = None,
+    owner = "Renault"
   )
 
-  private val episode1: HostedVideoPage = HostedVideoPage(
+  private val episode1WithoutNextPage: HostedVideoPage = HostedVideoPage(
+    sectionId = "renault-car-of-the-future",
     pageUrl = "https://www.theguardian.com/commercial/advertiser-content/renault-car-of-the-future/design-competition-episode1",
     pageName = episode1PageName,
     standfirst = "Renault challenged Central St Martins students to dream up the car of the future. The winning design will be announced at Clerkenwell Design Week (and on this site). Watch this short video to find out who made the shortlist",
     logoUrl = Static("images/commercial/logo_renault.jpg"),
-    zoeUrl = Static("images/commercial/ren_zoe.jpg"),
     bannerUrl = Static("images/commercial/ren_commercial_banner.jpg"),
     video = HostedVideo(
       mediaId = "renault-car-of-the-future",
@@ -45,15 +46,16 @@ object RenaultHostedPages {
       srcUrlOgg = "https://cdn.theguardian.tv/interactive/mp4/1080/2016/05/23/160523GlabsRenaultTestHD-3_hi.ogv",
       srcM3u8 = "https://cdn.theguardian.tv/interactive/2016/05/23/HLS/160523GlabsRenaultTestHD.m3u8"
     ),
-    nextPageName = episode2PageName
+    nextPage = None,
+    owner = "Renault"
   )
 
-  private val episode2: HostedVideoPage = HostedVideoPage(
+  private val episode2WithoutNextPage: HostedVideoPage = HostedVideoPage(
+    sectionId = "renault-car-of-the-future",
     pageUrl = "https://www.theguardian.com/commercial/advertiser-content/renault-car-of-the-future/design-competition-episode2",
     pageName = episode2PageName,
     standfirst = "A group of Central St Martins students took part in a competition to dream up the car of the future. The winning design is radical and intriguing. Meet the team whose blue-sky thinking may have created a blueprint for tomorrow's autonomous cars",
     logoUrl = Static("images/commercial/logo_renault.jpg"),
-    zoeUrl = Static("images/commercial/ren_zoe.jpg"),
     bannerUrl = Static("images/commercial/ren_commercial_banner.jpg"),
     video = HostedVideo(
       mediaId = "renault-car-of-the-future",
@@ -65,10 +67,15 @@ object RenaultHostedPages {
       srcUrlOgg = "https://cdn.theguardian.tv/interactive/mp4/1080/2016/06/03/160603GlabsRenaultTest3-3_hi.ogv",
       srcM3u8 = "https://cdn.theguardian.tv/interactive/2016/06/03/HLS/160603GlabsRenaultTest3.m3u8"
     ),
-    nextPageName = episode1PageName
+    nextPage = None,
+    owner = "Renault"
   )
 
-  lazy val defaultPage = teaser
+  private val teaser: HostedVideoPage = teaserWithoutNextPage.copy(nextPage = Some(episode1WithoutNextPage))
+
+  private val episode1: HostedVideoPage = episode1WithoutNextPage.copy(nextPage = Some(episode2WithoutNextPage))
+
+  private val episode2: HostedVideoPage = episode2WithoutNextPage.copy(nextPage = Some(episode1WithoutNextPage))
 
   def fromPageName(pageName: String): Option[HostedPage] = {
     pageName match {
